Volunteer beorn7 as 2.10 release shepherd
I'm willing to do this under the condition that I can run an
experiment with never freezing master. See addition in the file.

I believe, this way is more consistent with how we handle the bugfix
releases after the final minor release is cut. It's at least worth a
try.

Signed-off-by: beorn7 <bjoern@rabenste.in>