Cut v2.33.5
Note that this bugfix release also includes the update of the
containerd/containerd dependency, with the only purpose of shutting up
security scanner spam. We usually do not decument dependency updates
in the CHANGELOG. In this case, it isn't even doing anything. I plan
to mention the update in the announcement mail.

Signed-off-by: beorn7 <beorn@grafana.com>